+++ src/Presentation/Nop.Web/Nop.Web.csproj
@@ -296,6 +296,8 @@
     <Compile Include="Factories\IWidgetModelFactory.cs" />
     <Compile Include="Factories\IVendorModelFactory.cs" />
     <Compile Include="Factories\ITopicModelFactory.cs" />
+    <Compile Include="Factories\IShoppingCartModelFactory.cs" />
+    <Compile Include="Factories\ShoppingCartModelFactory.cs" />
     <Compile Include="Factories\TopicModelFactory.cs" />
     <Compile Include="Factories\VendorModelFactory.cs" />
     <Compile Include="Factories\WidgetModelFactory.cs" />
